Define static compartment update functions.
Within the Griffin model we must update state depending upon the compartment.
We have now defined the updates for S A U D T P compartments.
In doing so we have defined the determine_event function and the SAU_infection function.
We will remove branches within the simulation by assigning these functions to the update_ member variables of the PFalc state.