Source code for PFalc class.
The PFalc object can be initialised using an arbitrary disease state (compartment) and any immunity level.

Member function getIC and getIA allow for the return of the current immunity level.

The functions clearInfectionQueue, scheduleInfection and updateInfection control the queueing of future infections for the PFalc object.
This means we do not have to store any past data for the lagged force of infection.

We have provided functionality for the caching of the next infection within the PFalc object.
This has been done to minimise the indirection that is caused when searching for the infection_queue_.
It is expected that this will minimise the number of cache misses and therefore increase the efficiency of the simulation.