Define generalised function for a simulation.
A simulation, irrespective of what strain, should be able to follow a loop of while time is less than end time, run this function.
Therefore I have defined a templated simulation function.
The ability of this function to fully encapsulate everything we need is not yet fully established.
This will be established through the course of development.