clearInfectionQueue pops the underlying priority queue until size is one.
After some initial testing, we save around half a second by popping items off the queue until the size is one.
The alternative was to replace with an empty queue and add infinity to the queue.
The infinity is always in the queue so we never have to check if it is empty.

Remove cacheable infection.
This is a possible premature optimisation and has been done away with for now.