Wrap priority queue as an infection scheduler.
Instead of tracking information about the force of infection through time, we determine if a future infection will take place.
If it is true, we schedule that infection within this class.
We track both the time that the infection will occur and the amount of hypnozoites that it adds to the individual.

The current implementation does not clear infections.
This may be required when we start implementing treatment pathways.