Define how a single step of the Vivax model occurs.
Initial implementation of the vivax model.
There is an unnamed namespace with a global object kcached_data of type CachedVariables.
This object store information from the previous time step that is useful in speeding up calculations in the next.
Currently, we check if the object is up to date by checking the time that it was updated.

There are two variables to help with the exit of functions, kEXIT_INDIVIDUAL_LIVES and kEXIT_INDIVIDUAL_DIES.
These are to be used to ensure that true and false are not mixed up when determining if an individual is to live or die.

All compartments have their own update function that updates their state.

The one_step function is implemented at the bottom of the file.
The control flow is as follows:
	- Check that the cache is updated.
For each individual do the following:
	- Check if individual exceeds max age
	- Update individuals state.
	- Check if individual is birthing capable and add to immunity store if they are.
	- Return if individual is dead otherwise:
	- Update their age and omega value.
For each individual that has died, we birth the same number of individuals.
	- Cache the data for the next time step.