Change function signature and scaling factor
Function signature now expect constant eir.
The scaling factor for omega, now scales both omega and zeta.
I believe this will ensure that the number of bites is not scaled within the one_step function.

I think that the number of bites were scaled before because of the individual level heterogeneity not being scaled to 1.0.
The mosquitoes make a certain number of bites a day, and of the people that are alive that day, they should recieve those bites.
However zeta was not scaled so it could be larger number of bites.
I think that it was fine, but this is just an improvement.