Function traits implemented.
I found a link, https://functionalcpp.wordpress.com/2013/08/05/function-traits/ , that talks about function traits in c++.
This allows me to get the arity of a function.
I use this to split the parameter pack.
This will probably be adjusted through time as there are extra calculations that I do no require.