Create a population class for the human population.
The first implementation used a cache to store information about the
humans.
However, this cache was a global variable.
I was unsure of the functionality when using parallel simulation, but I
am quite certain it would not have worked.
As such I refactored the code so that each population will also store
the appropriate summary statistics for their simulation.
In the case of vivax this is the total of omega * zeta for the
population and the vector of maternal immunity.