Enabled emplace_back functionality in python for Population.
Could not manage to get this to work, but using a lambda fixes this
problem.
Note that it has to be a raw pointer and not a smart pointer or there is
ownership issues.
I could probably fix this with changing who owns the memory, but raw
pointer seems easy fix for now.